requirements.txt
pandas
pyyaml
openpyxl
langchain
python-dotenv
openai
tiktoken
faiss-cpu
fastapi
uvicorn
pytest
httpx
numexpr